switch to int p1

+++ src/CH/CH.hpp
@@ -1,7 +1,7 @@
 
-class CH : public CHTemplate<Edge, CHEdge, CHQueryEdge, float> {
+class CH : public CHTemplate<Edge, CHEdge, CHQueryEdge, int> {
 public:
-    CH(vector<vector<Edge>> inputGraph) : CHTemplate<Edge, CHEdge, CHQueryEdge, float>(inputGraph) {}
+    CH(vector<vector<Edge>> inputGraph) : CHTemplate<Edge, CHEdge, CHQueryEdge, int>(inputGraph) {}
 
     vector<int> contractVertex(int vertexNb, bool simulation) {
         int addedShortcuts = 0;
@@ -13,7 +13,7 @@ public:
         for (CHEdge ingoingEdge : ingoingEdges) {
             int u = ingoingEdge.getSourceVertex();
             vector<int> v_list;
-            vector<float> referenceWeight_list;
+            vector<int> referenceWeight_list;
             for (CHEdge outgoingEdge : outgoingEdges) {
                 int v = outgoingEdge.getDestinationVertex();
                 if (u != v) {
@@ -23,13 +23,13 @@ public:
             }
             if (v_list.size() > 0) {
                 // Compare path weight going through removed vertex x_i with shortest path weight without x_i
-                float stoppingDistance = *max_element(referenceWeight_list.begin(), referenceWeight_list.end());
-                pair<vector<float>, int> dikstraResult = dijkstraCH(u, v_list, stoppingDistance, vertexNb);
-                vector<float> shortest_list = dikstraResult.first;
+                int stoppingDistance = *max_element(referenceWeight_list.begin(), referenceWeight_list.end());
+                pair<vector<int>, int> dikstraResult = dijkstraCH(u, v_list, stoppingDistance, vertexNb);
+                vector<int> shortest_list = dikstraResult.first;
                 searchSpace += dikstraResult.second;
                 for (int i = 0; i<shortest_list.size(); i++) {
-                    float shortest = shortest_list[i];
-                    float referenceWeight = referenceWeight_list[i];
+                    int shortest = shortest_list[i];
+                    int referenceWeight = referenceWeight_list[i];
                     int v = v_list[i];
                     if ((shortest != -1) && (shortest <= referenceWeight)) {
                         // Found witness path : should not add shortcut
@@ -52,24 +52,24 @@ public:
         return vector<int>({addedShortcuts, searchSpace});
     }
 
-    pair<vector<float>, int> dijkstraCH(int s, vector<int> t_list, float stoppingDistance, int ignoreVertex) {
+    pair<vector<int>, int> dijkstraCH(int s, vector<int> t_list, int stoppingDistance, int ignoreVertex) {
         // Init
-        set<pair<float, int>> vertexSet;
-        vertexSet.insert(make_pair(0.0f, s));
+        set<pair<int, int>> vertexSet;
+        vertexSet.insert(make_pair(0, s));
         this->hops = this->hopsR; 
         this->vertexWeights = this->vertexWeightsR;
         vertexWeights[s] = 0;
 
-        vector<float> results = vector<float>(t_list.size(), -1);
+        vector<int> results = vector<int>(t_list.size(), -1);
         int visitedCount = 0;
         int searchSpace = 0;
 
         while (!vertexSet.empty()) {
             searchSpace++;
             // pop first vertex in the set
-            pair<float, int> visitedVertex = *(vertexSet.begin());
+            pair<int, int> visitedVertex = *(vertexSet.begin());
             vertexSet.erase(vertexSet.begin());
-            float visitedVertexWeight = visitedVertex.first;
+            int visitedVertexWeight = visitedVertex.first;
             int visitedVertexNb = visitedVertex.second;
 
             // update visited t count
@@ -84,16 +84,16 @@ public:
                 if (hops[visitedVertexNb] > hopLimits[hopIndex]) continue;
                 for (auto& e : g_R.getIncidenceList()[visitedVertexNb].first) {
                     if (e->getDestinationVertex() == ignoreVertex) continue;  // Ignore the vertex that will be removed in the graph
-                    float neighbourCurrentWeight = vertexWeights[e->getDestinationVertex()];
-                    float neighbourNewWeight = visitedVertexWeight + e->getWeight();
-                    if ((neighbourCurrentWeight == -1.0f) || (neighbourNewWeight < neighbourCurrentWeight)) {    // if smaller weight was found
-                        if (neighbourCurrentWeight != -1.0f) {                                                              
+                    int neighbourCurrentWeight = vertexWeights[e->getDestinationVertex()];
+                    int neighbourNewWeight = visitedVertexWeight + e->getWeight();
+                    if ((neighbourCurrentWeight == -1) || (neighbourNewWeight < neighbourCurrentWeight)) {    // if smaller weight was found
+                        if (neighbourCurrentWeight != -1) {                                                              
                             // if current weight not infinite : vertex already in queue : DELETE before inserting the updated vertex
-                            pair<float, int> currentNeighbour = make_pair(neighbourCurrentWeight, e->getDestinationVertex());
+                            pair<int, int> currentNeighbour = make_pair(neighbourCurrentWeight, e->getDestinationVertex());
                             vertexSet.erase(vertexSet.find(currentNeighbour));
                         }
                         // INSERT in queue
-                        pair<float, int> newNeighbour = make_pair(neighbourNewWeight, e->getDestinationVertex());
+                        pair<int, int> newNeighbour = make_pair(neighbourNewWeight, e->getDestinationVertex());
                         vertexSet.insert(newNeighbour);
                         // UPDATE weight
                         vertexWeights[e->getDestinationVertex()] = neighbourNewWeight;
@@ -105,7 +105,7 @@ public:
         // On-the-fly edge reduction
         for (int edgeIndex = 0; edgeIndex < g_R.getIncidenceList()[s].first.size(); edgeIndex++) {
             CHEdge* edge = g_R.getIncidenceList()[s].first[edgeIndex];
-            float newWeight = vertexWeights[edge->getDestinationVertex()];
+            int newWeight = vertexWeights[edge->getDestinationVertex()];
             if (newWeight != -1 && newWeight < edge->getWeight()) {
                 // Remove edge
                 g_R.removeEdge(edge);
@@ -120,7 +120,7 @@ public:
         return make_pair(results, searchSpace);
     }
 
-    vector<vector<CHQueryEdge>> convertToSearchGraph(CHGraph<Edge, CHEdge, float> g_H) {
+    vector<vector<CHQueryEdge>> convertToSearchGraph(CHGraph<Edge, CHEdge, int> g_H) {
         vector<vector<CHQueryEdge>> g_star = vector<vector<CHQueryEdge>>(this->n, vector<CHQueryEdge>());
         for (int u = 0; u<n; u++) {
             vector<CHEdge*> edgePtrs = g_H.getIncidenceList()[u].first;
