+++ src/CH/TCH.hpp
@@ -13,7 +13,7 @@ public:
         for (TCHEdge ingoingEdge : ingoingEdges) {
             int u = ingoingEdge.getSourceVertex();
             vector<int> v_list;
-            vector<float> referenceWeight_list;
+            vector<int> referenceWeight_list;
             vector<TTF> ttf_list;
             for (TCHEdge outgoingEdge : outgoingEdges) {
                 int v = outgoingEdge.getDestinationVertex();
@@ -25,13 +25,13 @@ public:
             }
             if (v_list.size() > 0) {
                 // Compare path weight going through removed vertex x_i with shortest path weight without x_i
-                float stoppingDistance = *max_element(referenceWeight_list.begin(), referenceWeight_list.end());
-                pair<vector<float>, int> dikstraResult = dijkstraCH(u, v_list, stoppingDistance, vertexNb);
-                vector<float> shortest_list = dikstraResult.first;
+                int stoppingDistance = *max_element(referenceWeight_list.begin(), referenceWeight_list.end());
+                pair<vector<int>, int> dikstraResult = dijkstraCH(u, v_list, stoppingDistance, vertexNb);
+                vector<int> shortest_list = dikstraResult.first;
                 searchSpace += dikstraResult.second;
                 for (int i = 0; i<shortest_list.size(); i++) {
-                    float shortest = shortest_list[i];
-                    float referenceWeight = referenceWeight_list[i];
+                    int shortest = shortest_list[i];
+                    int referenceWeight = referenceWeight_list[i];
                     TTF ttf = ttf_list[i];
                     int v = v_list[i];
                     if ((shortest != -1) && (shortest <= referenceWeight)) {
@@ -56,24 +56,24 @@ public:
         return vector<int>({addedShortcuts, searchSpace});
     }
 
-    pair<vector<float>, int> dijkstraCH(int s, vector<int> t_list, float stoppingDistance, int ignoreVertex) {
+    pair<vector<int>, int> dijkstraCH(int s, vector<int> t_list, int stoppingDistance, int ignoreVertex) {
         // Init
-        set<pair<float, int>> vertexSet;
-        vertexSet.insert(make_pair(0.0f, s));
+        set<pair<int, int>> vertexSet;
+        vertexSet.insert(make_pair(0, s));
         this->hops = this->hopsR; 
         this->vertexWeights = this->vertexWeightsR;
         vertexWeights[s] = 0;
 
-        vector<float> results = vector<float>(t_list.size(), -1);
+        vector<int> results = vector<int>(t_list.size(), -1);
         int visitedCount = 0;
         int searchSpace = 0;
 
         while (!vertexSet.empty()) {
             searchSpace++;
             // pop first vertex in the set
-            pair<float, int> visitedVertex = *(vertexSet.begin());
+            pair<int, int> visitedVertex = *(vertexSet.begin());
             vertexSet.erase(vertexSet.begin());
-            float visitedVertexWeight = visitedVertex.first;
+            int visitedVertexWeight = visitedVertex.first;
             int visitedVertexNb = visitedVertex.second;
 
             // update visited t count
@@ -88,16 +88,16 @@ public:
                 if (hops[visitedVertexNb] > hopLimits[hopIndex]) continue;
                 for (auto& e : g_R.getIncidenceList()[visitedVertexNb].first) {
                     if (e->getDestinationVertex() == ignoreVertex) continue;  // Ignore the vertex that will be removed in the graph
-                    float neighbourCurrentWeight = vertexWeights[e->getDestinationVertex()];
-                    float neighbourNewWeight = visitedVertexWeight + e->getMaxima();
-                    if ((neighbourCurrentWeight == -1.0f) || (neighbourNewWeight < neighbourCurrentWeight)) {    // if smaller weight was found
-                        if (neighbourCurrentWeight != -1.0f) {                                                              
+                    int neighbourCurrentWeight = vertexWeights[e->getDestinationVertex()];
+                    int neighbourNewWeight = visitedVertexWeight + e->getMaxima();
+                    if ((neighbourCurrentWeight == -1) || (neighbourNewWeight < neighbourCurrentWeight)) {    // if smaller weight was found
+                        if (neighbourCurrentWeight != -1) {                                                              
                             // if current weight not infinite : vertex already in queue : DELETE before inserting the updated vertex
-                            pair<float, int> currentNeighbour = make_pair(neighbourCurrentWeight, e->getDestinationVertex());
+                            pair<int, int> currentNeighbour = make_pair(neighbourCurrentWeight, e->getDestinationVertex());
                             vertexSet.erase(vertexSet.find(currentNeighbour));
                         }
                         // INSERT in queue
-                        pair<float, int> newNeighbour = make_pair(neighbourNewWeight, e->getDestinationVertex());
+                        pair<int, int> newNeighbour = make_pair(neighbourNewWeight, e->getDestinationVertex());
                         vertexSet.insert(newNeighbour);
                         // UPDATE weight
                         vertexWeights[e->getDestinationVertex()] = neighbourNewWeight;
